app/requirements.txt
geopandas
pandas
polars
numpy
python-arango
tqdm
shapely
scipy
networkx




